Create Goal Details List Api

+++ SyncingSyllabi.WebApi/SyncingSyllabi.Data/Models/Base/BaseTrackedModel.cs
@@ -7,8 +7,8 @@ namespace SyncingSyllabi.Data.Models.Base
     public class BaseTrackedModel : BaseModel
     {
         public DateTime DateCreated { get; set; }
-        public int CreatedBy { get; set; }
+        public Int64 CreatedBy { get; set; }
         public DateTime DateUpdated { get; set; }
-        public int UpdatedBy { get; set; }
+        public Int64 UpdatedBy { get; set; }
     }
 }
